fix(console): should be able to edit password and email in profile
