+++ packages/console/src/components/DelayedSuspenseFallback/index.module.scss
@@ -1,0 +1,6 @@
+.daisy {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+}
